Renamed i18nConsumer at withI18n
