+++ src/services/i18n/index.ts
@@ -1,5 +1,5 @@
 export { default as I18nProvider } from './view/I18nProvider/I18nProvider';
 export { ITranslateProps, ITranslateFunction, ITranslateKey } from './namespace';
-export { withI18n } from './view/i18nConsumer/i18nConsumer';
+export { withI18n } from './view/withI18n/withI18n';
 export { default as LanguageSelector } from './view/LanguageSelector/LanguageSelector';
 export { tKeys, I18nContext } from './constants';
